requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-groq
gspread
oauth2client
python-dotenv
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib